new OPTICAL_SIMULATION and update correspondingly

+++ Vanilla inverse solver/Codes/forward_solver/@FORWARD_SOLVER_CONVERGENT_BORN/FORWARD_SOLVER_CONVERGENT_BORN.m
@@ -51,8 +51,8 @@ classdef FORWARD_SOLVER_CONVERGENT_BORN < FORWARD_SOLVER
                 h.boundary_thickness = ones(1,3) * boundary_thickness;
             end
             assert(length(h.boundary_thickness) == 3, 'boundary_thickness should be either a 3-size vector or a scalar')
-            h.boundary_thickness_pixel = round((h.boundary_thickness*h.wavelength/abs(h.RI_bg))./(h.resolution.*2));
-            h.max_attenuation_width_pixel = round((attenuation_width*h.wavelength/abs(h.RI_bg))./(h.resolution.*2));
+            h.boundary_thickness_pixel = double(round((h.boundary_thickness*h.wavelength/abs(h.RI_bg))./(h.resolution.*2)));
+            h.max_attenuation_width_pixel = double(round((attenuation_width*h.wavelength/abs(h.RI_bg))./(h.resolution.*2)));
             if h.RI_xy_size(1)==0
                 h.RI_xy_size(1)=h.size(1);
             end
